[BUGFIX] Allow editing site settings without root page being present
If a site is configured for a non-existing root page id, an exception is
thrown when trying to edit site settings. Although the configured root
page should exist, especially during development it might be possible
that this is not the case (yet).

With this patch, this behavior is now hardened: A configured root page
is no longer expected to actually exist, and related meta information is
not populated to the docheader component in this case.

Resolves: #107419
Releases: main, 13.4
Change-Id: I29f7a7a23222265a43cc800c9634cd733098a515
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90652
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Benjamin Franzke <ben@bnf.dev>