[TASK] Remove main application services from failsafe container
Since #107536 we are able to serve the typo3 installer and maintenance
tool from the main index.php entrypoint, which means we can remove any
failsafe compatibility requirements from the main application classes,
allowing to use the full set of symfony DI features, without needing
applications to be defined in the failsafe container only to be
able to redirect to the install tool.

This mainly allows to inject services and use symfony DI, reducing
the amount of services that need to be manually wired in
service providers.

Releases: main
Resolves: #107574
Related: #107536
Change-Id: Ib456c2dc338e08dd0900c04ee99d433d5d355799
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90794
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>