[BUGFIX] Remove double encoding of LinkBrowser label output
The menu tabs of the LinkBrowser emitted labels by
using `htmlspecialchars()` in the assignment of
`AbstractLinkBrowserController->loadLinkHandlers()`.

This label gets extracted by `buildMenuArray()`. This method is
called in the `mainAction` of the link browser abstract, which
assigns that array to `{menuItems}` to be used in the Fluid template
`typo3/sysext/backend/Resources/Private/Layouts/LinkBrowser.html`.
There, the usual Fluid safeguards for escaping output
are in place.

By removing the extra `htmlspecialchars` characters like
a single quote (as reported in the issue) no longer get
mangled.

Resolves: #107531
Releases: main, 13.4
Change-Id: Ic1389ea6a9aae719fcf2460560b4ca78e4eb8934
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90804
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>