[BUGFIX] Align SplitButton usage to allow LinkButton
This patch now properly renders LinkButtons (`<a href...>`)
used within the "SplitButton" pane. Previously, any
kind of AbstractButton was rendered as a submit-button, which
led to wrong form submission and the inability to
actually utilize a link.

Also, `dataAttributes` are evaluated now for buttons added
into the SplitButton collection.

Resolves: #107336
Releases: main, 13.4
Change-Id: Ia53a7d756cf9899a9afb47794d5dfcf85484113b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91213
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>