[TASK] Improve panel rendering
The progress bar now renders its value if available and
hides it for regular users.

Panels can now use the progress bar element in addition to
.panel-progress. This allows more flexibility with a dynamically
updating component that is easier to use.

The panel variants for placeholder and hidden, previously used
only in FormEngine, are moved to the main scope. While they
currently share the same visuals, both are kept to preserve
semantic meaning in the class names.

Resolves: #107650
Releases: main, 13.4
Change-Id: I990cca2c31b1557a71dd270f66ff9db4ebed9041
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90989
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>