[!!!][TASK] Drop `prepend` and `append` TCA modes from Value Picker
We are removing the prepend and append modes from
the Value Picker used in FormEngina via TCA, as they
only cover a very niche use case and provide little
practical benefit.

There will be no direct replacement for these
options. If you still rely on them, we recommend
creating a custom input type that handles this
functionality and ensures proper accessibility
support.

Prepending or appending values to an existing
input is not an ideal pattern for input types,
as it often leads to unpredictable results and
inconsistent user experiences. It mixes
presentation and data logic, complicates input
validation, and can interfere with assistive
technologies. Instead, value selection should
focus on replacing or explicitly modifying the
input value in a controlled way.

In follow-up patches, we are reworking how
predefined values are selected by rebuilding the
Value Picker into new form and input types. The
existing replace mode will continue to be
supported, but the value picking mechanism will
evolve to provide a more consistent and accessible
experience.

Resolves: #107677
Releases: main
Change-Id: Ib54ff4cc111f070467f9c194ff857facc5f2372d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91023
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>