[BUGFIX] Prevent unintended operations in FormEngine
When FormEngine is loading or when it
gets reloaded, e.g. on changing the
record type, the doc header buttons
are now disabled to prevent unintended
operations, which previously could lead
to e.g. duplicated records.

Additionally, the `ShortcutButton` and
the `ButtonBarButton` do now also support
the disabled state functionality to
be in line with the other button types.

This is especially important since the
the shortcut button bar items were always
hidden behind the "ui-block" on form engine
initialization and therefore led to bad UX.

Now, buttons are properly disabled and visible,
providing better visual feedback to users during
the loading process.

Resolves: #107681
Releases: main, 13.4
Change-Id: Id4fc7e61ed887e7f3e7105d9565631ae317b89e5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91054
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Oli Bartsch <bo@cedev.de>