[BUGFIX] Maintain focus after clearing a clearable input field
When clicking the clear-indicator, the input value is cleared
and the indicator is removed from DOM and therefore focus
is moved back to `<body>`, which means that focus-flow is lost.

It is likely that a user that clears the value desires to change,
therefore a focus on the actual input should be the action
a user would expect.

This behavior also matches the one of the native type=search element.

The t3js-clearable class is removed from datetime element, since
this caused a double initialization of the clearable button (
date-picker already initializes the clearable handler for
the alternative datepicker input) and therefore prevented this
fix to operate.

Resolves: #107738
Releases: main, 13.4
Change-Id: Ie454c24f64063cd056b94449bb66370b255e5440
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91256
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>