[BUGFIX] Improve submodule link accessibility without breaking UX
The module name was recently added to the link
text (#107852) to provide better context for
accessibility ("Open <Module Name> module"
instead of just "Open module"). However, this
caused UX issues with long module names that
made the buttons visually odd.

This change moves the contextual information
to an `aria-label` attribute instead, which
provides the full context to screen readers
while keeping the visible link text short and
clean. This maintains accessibility improvements
without compromising the visual user experience.

For consistency, this change is also applied
to the modules overview in the About module.

Resolves: #107865
Related: #107852
Releases: main
Change-Id: I9407c87e4959dd9bf3783e03402f1ee9ad9d7aea
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91280
Tested-by: Jasmina Ließmann <typo3@minapok.eu>
Reviewed-by: Jasmina Ließmann <typo3@minapok.eu>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>