[BUGFIX] Restore password toggle and capslock warning functionality
The password toggle and capslock warning functionalities on the backend
user login form are no longer working since the recent login form
modifications with #107718. This is due to the removed `input-group`
containers, which are required when applying appropriate buttons for
said functionalities. These containers are now restored, re-enabling
both functionalities.

Resolves: #107876
Related: #107718
Releases: main, 13.4
Change-Id: I671fae95e149d8bfbe21baa3b4ad412b673a49f0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91410
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>