[BUGFIX] Improve module body scroll performance and reduce CPU spikes
This is basically the same fix as applied in #97144 for v10 and
in #93489 for v11 (and v12) some years ago.
The performance degraded in v13 due to our module refactoring
in #106381, as we moved the module overflow from `.module` to
`.module-body`.

The performance is not as bad as back in 2022, because
https://issues.chromium.org/issues/40369316 has been resolved
in the meantime (scrolling layers are automatically promoted to
a compositing layer now), but the chromes bugfix still requires a
composition between the transparent `.module-body` background and the
colored `.module` background. By applying an opaque background to the
scrolling element, we can ensure that the browser does not need to
compose against the background elements, which means it "only" has to
shift graphical buffers, without reapplying alpha-compositing (which
is faster than an entire rerender, but still not as fast as shifting
buffers).

Please see commit message of #97144 for more information on the
technical background.

Releases: main, 13.4
Resolves: #107921
Related: #106381
Related: #97144
Related: #93489
Change-Id: I73a686782c03e4c42891ae6670108819eb3cfdac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91348
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Garvin Hicking <garvin@hick.ing>