[TASK] Use system resource API to resolve resources in getFileAbsFileName
One of the main concepts of the system resources API is,
that resources are represented by objects. Object representation
is required to identify, distinguish and operate on resources
in a speaking and meaningful way in code.

Obtaining an absolute file system path from such resource violates
this concept, especially, when this string is manipulated or passed
around later on.

Despite this violation and in order to establish a more consistent state
regarding the possible usage of PKG: resources,
GeneralUtility::getFileAbsFileName() is changed to use the resource
API, instead of relying on the PackageManager only to resolve absolute
file system paths from a given resource identifier.

This mitigates the breakage in PageRenderer hooks, that now only
expose PKG: identifiers in internal properties, while these
properties are also delivered to hooks of third party extensions.

It also allows usage of the PKG: resource identifiers across the system,
including Fluid template paths.

While this improves the current state, until the core code base is ready
to work on resource objects only, obtaining the absolute path of a PKG:
resources is introduced and deprecated right away to communicate that this
concept violation will only exist for a short period of time and will be
removed until release of TYPO3 14LTS.

To allow the usage of the system resource API in getFileAbsFileName
the API is refactored in a way to have less duplicate dependencies
and is more concise in what responsibilities which code area has.

Specifically, the identifier handling is now responsible
for validity of the resource and having a dependency to package
management, whereas the resource objects themselves became more
lightweight.

Releases: main
Resolves: #107907
Related: #107537
Change-Id: I74dda714bba28a082ae60c105580dc581a286b07
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91464
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>