[BUGFIX] Prevent decoding of user supplied URLs in LinkBrowser
InputTransformer.toSearchParams for some reason decodes search
parameters after assembling them. This causes properly encoded URLs
to be changed, which violates RFC 1738 section
2.2 "All unsafe characters must always be encoded within a URL." [1]

We create an own URLSearchParams object as a workaround, in order
to avoid this faulty behavior, as this ensures that AjaxRequest
uses the values as-is.

The faulty behavior will be fixed separated for v14 only, see #108052.
This workaround is scoped to the reported issue. A followup patch will
address the decoding inside `AjaxRequest(...).withQueryArguments`,
see #108052.

[1] https://www.rfc-editor.org/rfc/rfc1738#section-2.2

Resolves: #107475
Related: #108052
Releases: main, 13.4
Change-Id: I046eda901a8fbd64d227a67d67ca60e13ef05828
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91613
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>