[TASK] Add "debug:backend:modules", "debug:backend:routes" commands
These CLI commands help developers to see:

- the list of registered backend modules and their hierarchy, along
  their extension origin and defined (translated) labels

- the list of backend routes and their targets

Two backend module list metadata items are not exposed through the
ModuleInterface API (and normally not relevant), so they are fetched
on a low-level basis.

With these command, developers are able to:

- debug the hierarchy and make choices about possible naming
  conflicts, ill-fated identifiers and proper ordering

- investigate AJAX routes, possible cross-concerns and scopes

Since this is regarded internal usage, it is not exposed
as a feature and uses a "debug:" CLI scope.

Resolves: #108057
Resolves: #107830
Related: #107790
Related: #108047
Related: #108045
Related: #108048
Releases: main
Change-Id: If1cabc035fa3c167d8c7a6ee571609ad5ccd1b69
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91566
Tested-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>