[BUGFIX] Prevent tree node highlights splitting encoded characters
This change rewrites the createNodeContentLabel method, avoiding
nodes purely being created for HTML escapes and simple replaces
for a search, as well as final "innerHTML" assembly of the node
content.

We utilize the same regex as before, in addition to a capture group,
but now with splitting instead of replacing. The capture group
retains the split-separator in the resulting matches, allowing us
to simply map through all results, wrapping every second one with
the node highlight span.

This results in cleaner code and subquently fixes bugs like the
previous implementation splitting HTML entities when "&" were
searched for.

Resolves: #108003
Releases: main, 13.4
Change-Id: Ic9202633a4da0ceb11881608f881d35ad0c82b4f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91590
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>