[TASK] Update install tool module identifiers
After moving the Install Tool related modules
to the System top-level module in #108069, their
identifiers are now updated accordingly. Also
their routing is done manually, why it is
necessary to update the path in the router
JS component as well.

Although we usually avoid prefixing submodule
identifiers with their top-level identifier,
it is necessary here to ensure clear separation,
as the names would be too generic.

Additionally, the main title of the Database
Integrity module has been changed to “Database”
for better clarity and to align with recent
single-word title renamings, see #107628.

Resolves: #108094
Related: #108069
Related: #107628
Releases: main
Change-Id: Ic6739691b80d6e6854fbbb351fef4e00ed081a7c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91610
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>