[FEATURE] Add attribute AsNonSchedulableCommand for "schedulable"
With #101567 the usage of Symfony's `#[AsCommand]` attribute
has been introduced, which allows to configure a CLI Symfony Command
name and description (and some other options).

It did not support TYPO3's custom implementation for the `schedulable`
option, which allows to flag a CLI command to not be able to be
scheduled in EXT:scheduler, because a decision was made to use
the native Symfony attribute instead of introducing a custom
variant just for these scenarios. This was made under the assumption
that only rarely commands would need to be set to "non schedulable".

However, recent debugging command additions revealed that this choice
leads to scattered configuration and gets uilized in more cases.

This required to list such a command with `schedulable: false` in the
`Services.yaml` / `Services.php` definition.

To improve this situation, the PHP attribute
`\TYPO3\CMS\Core\Attribute\AsNonSchedulableCommand` has
been created. Any Symfony Command can receive this attribute,
so that the automatic Scheduler registry will ignore commands
with this implementation.

Another upside of this is that an IDE like PhpStorm is capable
of showing all usages of that attribute inside a project.

All TYPO3 Core commands are adjusted to receive the attribute,
and remove the Service definition.

Resolves: #107151
Related: #101567
Releases: main
Change-Id: I3e620a6060aff40eec8ef5e56c90fb305cbf405c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91623
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>