[FEATURE] Add QR Codes module
A new backend module called "QR Codes" has been added.
The "Redirects" and "QR Codes" module are now submodules
of "Link Management".

The new module leverages the existing logic of the
"Redirects" but uses a custom TCA type to adapt the GUI
for the QR Codes.

It enables the user to create and download a QR Code in
different sizes and formats (PNG, SVG).

Resolves: #107756
Releases: main
Change-Id: Ic19519e63c168b3ba15fe3795d04454160d7e975
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91655
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>