[TASK] Migrate remaining commands to AsCommand attribute
Migrate a couple of missed commands to use the `#[AsCommand]`
and `#[AsNonSchedulableCommand]` attributes for registration.

Migrate `SendEmailCommand` to fully use the `#[AsCommand]`
attribute. A definition for a `SendEmailCommand` alias was
incorrectly kept. Aliases are defined as third argument of
the `#[AsCommand]` attribute.

Additionally, the feature rst is improved and a code
example is added.

Resolves: #108203
Related: #107151
Releases: main
Change-Id: Iaac6b4ab8ceafd768f4937ecdfbd1492fa48ac29
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91760
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>