[TASK] Rename “Page” module to “Layout”
The module previously named “Page” is renamed
to “Layout” to better communicate its actual
purpose. The module focuses on managing the
layout of a page, arranging content elements
in the correct columns, and maintaining the
structural composition of the site.

Note: This patch does not change the module
identifier, since this has a couple of side
effects which need to be handled carefully.

Resolves: #108233
Related: #107628
Releases: main
Change-Id: Ic3fcbff0cbef89b4d77d33a0a653522a3db58c92
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91792
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Lina Wolf <112@linawolf.de>