[TASK] Add single submodule promotion
Adds a internal appearance option for top-level modules
that automatically promotes a single submodule to
standalone status.

When a parent module has only one submodule and this
option is enabled, the submodule becomes a standalone
top-level module (like dashboard), inheriting the
parent's navigation component, position, and aliases.

This is applied to the "media" module, so when
"media_management" is the only submodule, like in
99% of all installations, it appears as a standalone
module in the menu.

Resolves: #108235
Related: #107628
Releases: main
Change-Id: Ie28e2562458f6c6adc508f58ef44af6754eef49e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91796
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>