[BUGFIX] Preserve focused translation when switching to single-language view
When switching from a multi-language view to a
single-language view (e.g. in Page or View module),
the system now correctly preserves the user's
focused translation.

Previously, `getPrimaryLanguageId()` always returned the
first element of selectedLanguageIds, which was mostly
the default as this is automatically added in comparison
views, ignoring which translation the user was actually
working with.

The new logic in PageContext::getPrimaryLanguageId():

- If exactly 1 non-default language is selected => use that translation
- If 0 or 2+ non-default languages are selected => use default (0)

This matches user expectations: when they have
default + one translation selected in comparison
view, switching to single-language view should show
that translation, not fall back to the default.

Resolves: #108245
Releases: main
Change-Id: Idf17df5b17141d911f9b983d3a9833a088369b0e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91809
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Sven Liebert <mail@sven-liebert.de>
Tested-by: Sven Liebert <mail@sven-liebert.de>