Revert "[TASK] Harmonize label positioning in element browser"
This reverts commit 018a9f761721b20690a0589b5da350223c146937.

Reason for revert:

This patch (91790) unintentionally moves the element browser layout
backwards in terms of the architecture we are actively building. It
reintroduces static Bootstrap column classes that we are systematically
removing, and it reduces the available space for essential information.
For example, email fields may become truncated, and users are forced to
scroll to reach the “Set” button. In several element browser views, the
new layout also creates excessive unused whitespace.

The previous solution was deliberately designed to accommodate the space
requirements of the element browser and used dedicated markup because
this view has different needs than other form areas. The patch removed
that specialization, but without actually resolving the underlying issue
it was meant to address.

The original problem is a side effect of the form styling rebuild
introduced in v13. Therefore, any real fix must also be backported to
TYPO3 v13. Patch 91790 does not resolve the issue for current users.

A minimal, targeted replacement patch (91821) has been created to
correctly address the label-rendering problem. It fixes the broken
ellipsis on form labels, restores the intended layout, and increases
the available label space, especially important for languages like
German, while remaining fully compatible with both v13 and main.

We are therefore reverting 91790 in favor of 91821.

Resolves: #108273
Reverts: #107985
Releases: main

Change-Id: I367c16807f6f70aa9ef4c2f37b2932c8ee0e625e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91832
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Tested-by: Josef Glatz <typo3@josefglatz.at>
Reviewed-by: Oli Bartsch <bo@cedev.de>