[TASK] Avoid invalid dropdown button in module menu processing
Previously, the menu was rendered as a <select>,
which meant the first or currently selected <option>
acted as the visible label. The menu itself had no
separate label.

As a fallback, we now use the first menu item title
as the button label, ensuring the DropDownButton is
valid. The button will still always display the active
item, because setShowActiveLabelText(true) is set.

Resolves: #108284
Related: #107875
Releases: main
Change-Id: Ieacd672f65f1fc3942e22a68b6cc0e006fe4217f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91843
Reviewed-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Oli Bartsch <bo@cedev.de>