[BUGFIX] Do not render "Create new record" button in single table view
In single table view, it is unexpected to get the
general "Create new record" button in the doc header
displayed. Especially because it allows to create
records, which are not visible after save.

Instead, the specific "create new <Table title>"
button of the selected table is now displayed.

Resolves: #108498
Releases: main, 13.4
Change-Id: I86e6c5ee4aa057dbd68c9382bc4470d5b7d1b880
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92134
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oli Bartsch <bo@cedev.de>
Reviewed-by: Oli Bartsch <bo@cedev.de>