[TASK] Centralize and streamline open documents functionality
This change refactors the "open documents" functionality
to be more consistent after more than 20 years of being "just there".

Open Documents = documents that are opened by FormEngine of the current
users without having explicitly saved or closed, are stored in the
backend users' session storage "uc".

Previously, open documents were identified by an MD5 hash of the entire
edit session parameters. This meant the same record could appear multiple
times in the list if opened with different parameters, and multiple records
edited together would only show as a single entry.

This way, TYPO3 Backend now tracks open documents properly:

- Only one entry is added even if a record is edited multiple times
- New records are tracked as well and can be utilized

In addition, instead of a numeric array list, readonly
OpenDocument objects are used so they can be of use
in multiple places later-on as well as a better separation
of concerns with repository pattern.

Next to extracting a lot the logic out of EditDocumentController,
this change allows to eventually change the storage format
to a native database table at a later stage.

Resolves: #108583
Releases: main
Change-Id: Ic2f2f23a6766a1d3b49848ea0bff7b514a4e926c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/92233
Tested-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>